Make it easier to work with documentation comments
Enter will now automatically continue documentation comments with the correct 'indentation' for subsequent lines. Since it would be annoying to have to fix the indentation manually when starting a new section, that is done automatically too when you type the beginning of a new section, e.g. 'Returns:' or 'Throws MeaningfulError:'.